receiver/requirements.txt
connexion[flask,uvicorn,swagger-ui]
PyYAML
kafka-python